+++ CHANGELOG.md
@@ -3,6 +3,15 @@ MATRIXIO Python MALOS Driver project adheres to [Semantic Versioning](http://sem
 
 ## [Unreleased]
 
+## [0.4.1] - 2019-04-24
+### Changed
+- Updated pyzmq dependency from version 17.1.2 to 18.0.1.
+
+## [0.4.0] - 2019-04-23
+### Added
+- `get_frame` asynchronous generator to handle
+  video streaming from Malos.
+
 ## [0.3.2] - 2018-11-14
 ### Changed
 - Set matrix-io-proto to be at least version 0.0.29